+++ src/builder.ts
@@ -6,7 +6,7 @@
 
 import { EsbuildBuilder } from './builder/esbuild'
 import { WebpackBuilder } from './builder/webpack'
-import { ElectronEsbuildConfigItem } from './config/types'
+import { ElectronEsbuildConfigItem, ItemConfig, PossibleConfiguration } from './config/types'
 import { isEsbuild, isWebpack } from './config/utils'
 import { unsupportedType } from './console'
 
@@ -18,9 +18,9 @@ export interface Builder {
 }
 
 export function createBuilders(
-  mainConfig: ElectronEsbuildConfigItem,
-  rendererConfig: ElectronEsbuildConfigItem,
-): [Builder, Builder] {
+  mainConfig: ElectronEsbuildConfigItem<PossibleConfiguration, ItemConfig>,
+  rendererConfig: ElectronEsbuildConfigItem<PossibleConfiguration | null>,
+): [Builder, Builder | null] {
   let mainBuilder: Builder
   let rendererBuilder: Builder
 
@@ -32,7 +32,9 @@ export function createBuilders(
     unsupportedType(mainConfig.fileConfig.type, 'main')
   }
 
-  if (isEsbuild(rendererConfig)) {
+  if (rendererConfig === null || rendererConfig.fileConfig === null) {
+    return [mainBuilder, null]
+  } else if (isEsbuild(rendererConfig)) {
     rendererBuilder = new EsbuildBuilder(rendererConfig)
   } else if (isWebpack(rendererConfig)) {
     rendererBuilder = new WebpackBuilder(rendererConfig)